rework new settings and tests for clarity
