+++ bookwyrm/middleware/__init__.py
@@ -4,4 +4,5 @@ from .timezone_middleware import TimezoneMiddleware
 from .ip_middleware import IPBlocklistMiddleware
 from .file_too_big import FileTooBig
 from .force_logout import ForceLogoutMiddleware
-from .bookwyrm_security import BookWyrmSecurityChecks
+from .require_signed_get import RequireSignedGet
+from .require_login import RequireLoginNearlyEverywhere
