+++ bookwyrm/settings.py
@@ -118,7 +118,8 @@ MIDDLEWARE = [
     "django.middleware.csrf.CsrfViewMiddleware",
     "csp.middleware.CSPMiddleware",
     "django.contrib.auth.middleware.AuthenticationMiddleware",
-    "bookwyrm.middleware.BookWyrmSecurityChecks",
+    "bookwyrm.middleware.RequireSignedGet",
+    "bookwyrm.middleware.RequireLoginNearlyEverywhere",
     "bookwyrm.middleware.TimezoneMiddleware",
     "bookwyrm.middleware.IPBlocklistMiddleware",
     "django.contrib.messages.middleware.MessageMiddleware",
