+++ bookwyrm/templates/landing/landing.html
@@ -4,7 +4,7 @@
 {% load landing_page_tags %}
 
 {% block panel %}
-{% if not require_login_everywhere %}
+{% if not require_login_nearly_everywhere %}
 <div class="block is-hidden-tablet">
     <h2 class="title has-text-centered">{% trans "Recent Books" %}</h2>
 </div>
