+++ bookwyrm/templates/settings/federation/settings.html
@@ -44,7 +44,7 @@
         </div>
         <div class="field">
             <label class="label" for="id_require_login_everywhere">
-                {{ form.require_login_everywhere }}
+                {{ form.require_login_nearly_everywhere }}
                 {% trans "Prevent unauthenticated views" %}
             </label>
             <span class="help" id="desc_require_login_everywhere">
