+++ bookwyrm/tests/middleware/test_get_signatures.py
@@ -1,6 +1,5 @@
-"""test security middleware"""
+"""test get signatures middleware"""
 
-from unittest.mock import patch
 from collections import namedtuple
 import json
 import pathlib
@@ -15,155 +14,6 @@ from bookwyrm import models
 from bookwyrm.activitypub.base_activity import get_representative
 from bookwyrm.settings import DOMAIN
 from bookwyrm.signatures import make_signature, create_key_pair
-
-
-class TestBookWyrmSecurityChecks(TestCase):
-    """lets get fuzzing"""
-
-    @classmethod
-    def setUpTestData(cls):
-        """create users and test data"""
-
-        cls.user = models.User.objects.create_user(
-            f"mouse@{DOMAIN}",
-            "mouse@example.com",
-            "changeme",
-            local=True,
-            localname="mouse",
-        )
-
-        cls.site = models.SiteSettings.get()
-        cls.site.require_login_everywhere = False
-        cls.site.block_incoming_search = False
-        cls.site.save(
-            update_fields=["block_incoming_search", "require_login_everywhere"]
-        )
-
-    @override_settings(
-        MIDDLEWARE=[
-            "django.contrib.sessions.middleware.SessionMiddleware",
-            "django.contrib.auth.middleware.AuthenticationMiddleware",
-            "bookwyrm.middleware.BookWyrmSecurityChecks",
-        ]
-    )
-    def test_require_login_everywhere(self):
-        """block pages with require_login_everywhere turned on"""
-
-        self.client.user = AnonymousUser()
-
-        # default is to allow user pages
-        response = self.client.get("/user/mouse")
-        self.assertEqual(response.status_code, 200)
-
-        # turn on require_login_everywhere
-        self.site.require_login_everywhere = True
-        self.site.save(update_fields=["require_login_everywhere"])
-
-        # should redirect to login for user page
-        response = self.client.get("/user/mouse")
-        self.assertEqual(response.status_code, 302)
-        self.assertEqual(response.url, "/login/?next=/user/mouse")
-
-        with patch(
-            "bookwyrm.activitystreams.ActivityStream.get_activity_stream"
-        ) as mock:
-            response = self.client.get("/discover")
-            self.assertEqual(response.status_code, 302)
-
-        response = self.client.get("")
-        self.assertEqual(response.status_code, 200)
-
-    @override_settings(MIDDLEWARE=["bookwyrm.middleware.BookWyrmSecurityChecks"])
-    def test_require_login_everywhere_allowed_pages(self):
-        """don't block allowlist"""
-
-        # should allow allow_list pages
-        response = self.client.get("/robots.txt")
-        self.assertEqual(response.status_code, 200)
-
-        response = self.client.get("/2fa-check")
-        self.assertEqual(response.status_code, 200)
-
-    @override_settings(
-        MIDDLEWARE=[
-            "django.contrib.sessions.middleware.SessionMiddleware",
-            "django.contrib.auth.middleware.AuthenticationMiddleware",
-            "bookwyrm.middleware.BookWyrmSecurityChecks",
-        ]
-    )
-    def test_require_login_everywhere_logged_in(self):
-        """allow logged in users"""
-
-        # turn on require_login_everywhere
-        self.site.require_login_everywhere = True
-        self.site.save(update_fields=["require_login_everywhere"])
-
-        self.client.user = self.user
-        self.client.login(username=self.user.username, password="changeme")
-
-        response = self.client.get("/user/mouse")
-        self.assertEqual(response.status_code, 200)
-
-        with patch(
-            "bookwyrm.activitystreams.ActivityStream.get_activity_stream"
-        ) as mock:
-            response = self.client.get("/discover")
-            self.assertEqual(response.status_code, 200)
-
-    @override_settings(
-        MIDDLEWARE=[
-            "django.contrib.sessions.middleware.SessionMiddleware",
-            "django.contrib.auth.middleware.AuthenticationMiddleware",
-            "bookwyrm.middleware.BookWyrmSecurityChecks",
-        ]
-    )
-    def test_require_login_everywhere_api_request(self):
-        """allow api requests"""
-
-        self.client.user = AnonymousUser()
-
-        # default is to allow user pages
-        response = self.client.get("/user/mouse")
-        self.assertEqual(response.status_code, 200)
-
-        # turn on require_login_everywhere
-        self.site.require_login_everywhere = True
-        self.site.save(update_fields=["require_login_everywhere"])
-
-        # allow API requewsts
-        response = self.client.get(
-            "/user/mouse",
-            headers={
-                "Accept": 'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-            },
-        )
-        self.assertEqual(response.status_code, 200)
-
-    @override_settings(MIDDLEWARE=["bookwyrm.middleware.BookWyrmSecurityChecks"])
-    def test_block_incoming_search(self):
-        """disallow search endpoint"""
-
-        # TODO change this to an API request with headers
-        response = self.client.get(
-            "/search/?q=beep",
-            headers={
-                "Host": DOMAIN,
-                "Accept": 'application/ld+json; profile="https://www.w3.org/ns/activitystreams"',
-            },
-        )
-        self.assertEqual(response.status_code, 200)
-
-        self.site.block_incoming_search = True
-        self.site.save(update_fields=["block_incoming_search"])
-
-        response = self.client.get(
-            "/search/?q=boop",
-            headers={
-                "Host": DOMAIN,
-                "Accept": 'application/ld+json; profile="https://www.w3.org/ns/activitystreams"',
-            },
-        )
-        self.assertEqual(response.status_code, 403)
 
 
 class TestBookWyrmGetSignatures(TestCase):
@@ -195,7 +45,7 @@ class TestBookWyrmGetSignatures(TestCase):
         MIDDLEWARE=[
             "django.contrib.sessions.middleware.SessionMiddleware",
             "django.contrib.auth.middleware.AuthenticationMiddleware",
-            "bookwyrm.middleware.BookWyrmSecurityChecks",
+            "bookwyrm.middleware.RequireSignedGet",
         ]
     )
     def test_get_signature_not_required(self):
@@ -220,7 +70,7 @@ class TestBookWyrmGetSignatures(TestCase):
         MIDDLEWARE=[
             "django.contrib.sessions.middleware.SessionMiddleware",
             "django.contrib.auth.middleware.AuthenticationMiddleware",
-            "bookwyrm.middleware.BookWyrmSecurityChecks",
+            "bookwyrm.middleware.RequireSignedGet",
         ]
     )
     @responses.activate
@@ -261,7 +111,7 @@ class TestBookWyrmGetSignatures(TestCase):
         MIDDLEWARE=[
             "django.contrib.sessions.middleware.SessionMiddleware",
             "django.contrib.auth.middleware.AuthenticationMiddleware",
-            "bookwyrm.middleware.BookWyrmSecurityChecks",
+            "bookwyrm.middleware.RequireSignedGet",
         ]
     )
     @responses.activate
@@ -314,7 +164,7 @@ class TestBookWyrmGetSignatures(TestCase):
         MIDDLEWARE=[
             "django.contrib.sessions.middleware.SessionMiddleware",
             "django.contrib.auth.middleware.AuthenticationMiddleware",
-            "bookwyrm.middleware.BookWyrmSecurityChecks",
+            "bookwyrm.middleware.RequireSignedGet",
         ]
     )
     @responses.activate
@@ -367,7 +217,8 @@ class TestBookWyrmGetSignatures(TestCase):
         MIDDLEWARE=[
             "django.contrib.sessions.middleware.SessionMiddleware",
             "django.contrib.auth.middleware.AuthenticationMiddleware",
-            "bookwyrm.middleware.BookWyrmSecurityChecks",
+            "bookwyrm.middleware.RequireSignedGet",
+            "bookwyrm.middleware.RequireLoginNearlyEverywhere",
         ]
     )
     @responses.activate
@@ -387,8 +238,10 @@ class TestBookWyrmGetSignatures(TestCase):
         responses.add(responses.GET, self.sender.remote_id, json=data, status=200)
 
         self.site.require_signed_get = True
-        self.site.require_login_everywhere = True
-        self.site.save(update_fields=["require_signed_get", "require_login_everywhere"])
+        self.site.require_login_nearly_everywhere = True
+        self.site.save(
+            update_fields=["require_signed_get", "require_login_nearly_everywhere"]
+        )
 
         # no signed headers, go away
         response = self.client.get("/user/mouse")
