+++ bookwyrm/tests/views/test_search.py
@@ -11,7 +11,7 @@ from django.test.client import RequestFactory
 
 from bookwyrm import models, views
 from bookwyrm.book_search import SearchResult
-from bookwyrm.settings import BASE_URL
+from bookwyrm.settings import BASE_URL, DOMAIN
 from bookwyrm.tests.validate_html import validate_html
 
 
@@ -40,6 +40,8 @@ class Views(TestCase):
             remote_id="https://example.com/book/1",
             parent_work=cls.work,
         )
+
+        cls.site = models.SiteSettings.get()
 
     def setUp(self):
         """individual test setup"""
@@ -224,3 +226,27 @@ class Views(TestCase):
         self.assertIsInstance(response, TemplateResponse)
         validate_html(response.render())
         self.assertEqual(response.context_data["results"][0], booklist)
+
+    def test_block_incoming_search(self):
+        """disallow search endpoint"""
+
+        response = self.client.get(
+            "/search/?q=beep",
+            headers={
+                "Host": DOMAIN,
+                "Accept": 'application/ld+json; profile="https://www.w3.org/ns/activitystreams"',
+            },
+        )
+        self.assertEqual(response.status_code, 200)
+
+        self.site.block_incoming_search = True
+        self.site.save(update_fields=["block_incoming_search"])
+
+        response = self.client.get(
+            "/search/?q=boop",
+            headers={
+                "Host": DOMAIN,
+                "Accept": 'application/ld+json; profile="https://www.w3.org/ns/activitystreams"',
+            },
+        )
+        self.assertEqual(response.status_code, 403)
